cloud_backend/requirements.txt
# Cloud Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0
pymongo==4.6.0
redis==5.0.1
boto3==1.34.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
celery==5.3.4
pytest==7.4.3
pytest-asyncio==0.21.1
hypothesis==6.92.1
python-dotenv==1.0.0
cryptography>=42.0.0
google-cloud-speech==2.21.0
openai==1.3.7